Split render shader and pass caches
