+++ src/render/wgpu/render_light_item.rs
@@ -12,8 +12,8 @@ use super::material::RenderUniformValue;
 use super::render_item::LightRenderItem;
 use super::render_item::LinesRenderItem;
 use super::render_item::RenderItem;
-use super::render_item::create_render_pass;
 use super::render_item::get_color;
+use super::render_pass_cache::create_render_pass;
 use super::render_resource::RenderResourceManager;
 use super::texture::RenderTexture;
 use crate::conversion::light_shape::create_light_shape;
