+++ src/render/wgpu/render_mesh_item.rs
@@ -5,11 +5,11 @@ use super::material::RenderUniformValue;
 use super::mesh::RenderMesh;
 use super::render_item::MeshRenderItem;
 use super::render_item::RenderItem;
-use super::render_item::create_render_pass;
 use super::render_item::get_bool;
 use super::render_item::get_color;
 use super::render_item::get_float;
 use super::render_item::get_texture;
+use super::render_pass_cache::create_render_pass;
 use super::render_resource::RenderResourceManager;
 use crate::model::scene::Light;
 use crate::model::scene::Material;
